EMP-553: refactored crm14 and crm15 into their own model package
