+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/model/crm14/crm15/Crm15PropertyModel.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.equinity.historicaldata.model;
+package uk.gov.justice.laa.crime.equinity.historicaldata.model.crm14.crm15;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -10,7 +10,7 @@ import java.util.List;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Crm15InvestmentsModel {
+public class Crm15PropertyModel {
     @JsonProperty("row")
-    public List<Crm15InvestmentDetailsModel> investments;
+    public List<Crm15PropertyDetailsModel> properties;
 }
